Move override_gentoo.py to _internal
