+++ certbot-apache/certbot_apache/entrypoint.py
@@ -11,7 +11,7 @@ from certbot_apache import override_fedora
 from certbot_apache import override_darwin
 from certbot_apache import override_debian
 from certbot_apache._internal import override_centos
-from certbot_apache import override_gentoo
+from certbot_apache._internal import override_gentoo
 from certbot_apache import override_suse
 
 OVERRIDE_CLASSES = {
