+++ certbot-apache/certbot_apache/tests/gentoo_test.py
@@ -8,7 +8,7 @@ from certbot.compat import filesystem
 from certbot.compat import os
 
 from certbot_apache import obj
-from certbot_apache import override_gentoo
+from certbot_apache._internal import override_gentoo
 from certbot_apache.tests import util
 
 
@@ -52,7 +52,7 @@ class MultipleVhostsTestGentoo(util.ApacheTest):
                                                     config_root=config_root,
                                                     vhost_root=vhost_root)
 
-        with mock.patch("certbot_apache.override_gentoo.GentooParser.update_runtime_variables"):
+        with mock.patch("certbot_apache._internal.override_gentoo.GentooParser.update_runtime_variables"):
             self.config = util.get_apache_configurator(
                 self.config_path, self.vhost_path, self.config_dir, self.work_dir,
                 os_info="gentoo")
@@ -85,7 +85,7 @@ class MultipleVhostsTestGentoo(util.ApacheTest):
         self.config.parser.apacheconfig_filep = filesystem.realpath(
             os.path.join(self.config.parser.root, "../conf.d/apache2"))
         self.config.parser.variables = {}
-        with mock.patch("certbot_apache.override_gentoo.GentooParser.update_modules"):
+        with mock.patch("certbot_apache._internal.override_gentoo.GentooParser.update_modules"):
             self.config.parser.update_runtime_variables()
         for define in defines:
             self.assertTrue(define in self.config.parser.variables.keys())
@@ -95,7 +95,7 @@ class MultipleVhostsTestGentoo(util.ApacheTest):
         """Make sure we don't call binary dumps other than modules from Apache
         as this is not supported in Gentoo currently"""
 
-        with mock.patch("certbot_apache.override_gentoo.GentooParser.update_modules"):
+        with mock.patch("certbot_apache._internal.override_gentoo.GentooParser.update_modules"):
             self.config.parser.update_runtime_variables()
             self.config.parser.reset_modules()
         self.assertFalse(mock_subprocess.called)
